+++ bff/src/main/java/bg/com/bo/bff/providers/dtos/request/crypto/currency/CryptoCurrencyExchangeOperationRequest.java
@@ -6,8 +6,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @Builder
 @AllArgsConstructor
@@ -15,7 +13,7 @@ import java.math.BigDecimal;
 public class CryptoCurrencyExchangeOperationRequest {
 
     @JsonProperty("importe")
-    private BigDecimal amount;
+    private Double amount;
 
     @JsonProperty("codMoneda")
     private Integer currencyCode;
